tt-01-django-sql/requirements.txt
Django==3.1
djangorestframework==3.12.2
Pillow==8.1.2
django-crispy-forms==1.9.2
whitenoise==5.2.0
gunicorn
django-heroku